+++ client/src/app/admin-dashboard/components/studies-control/new-study/tag-field.component.ts
@@ -10,6 +10,10 @@ import {
   VideoOptionUpload,
   videoOptionUploadRendererTester,
 } from '../../../../shared/components/custom-fields/video-option-upload.component';
+import {
+  OneOfField,
+  oneOfFieldTester,
+} from '../../../../shared/components/custom-fields/one-of.component';
 
 @Component({
   selector: 'tag-field',
@@ -36,6 +40,7 @@ export class TagFieldComponent implements OnInit {
     ...angularMaterialRenderers,
     { tester: fileListControlRendererTester, renderer: FileListField },
     { tester: videoOptionUploadRendererTester, renderer: VideoOptionUpload },
+    { tester: oneOfFieldTester, renderer: OneOfField },
   ];
   data: any = {};
 
@@ -51,7 +56,11 @@ export class TagFieldComponent implements OnInit {
   @Input() field: TagField;
 
   ngOnInit(): void {
-    this.schema = this.field.getDataSchema();
+    this.loadSchema();
+  }
+
+  async loadSchema() {
+    this.schema = await this.field.getDataSchema();
     this.uischema = this.field.getUISchema();
   }