+++ client/src/app/tagging-interface/tagging-interface.module.ts
@@ -7,10 +7,11 @@ import { SharedModule } from '../shared/shared.module';
 import { TaggingInterface } from './components/tagging-interface.component';
 import { TaggingForm } from './components/tagging-form.compont';
 import { TaggingLanding } from './components/tagging-landing.component';
+import { VideoRecordingModule } from '../video-recording/video-recording.module';
 
 @NgModule({
   declarations: [TaggingInterface, TaggingLanding, TaggingForm],
-  imports: [TaggingInterfaceRoutingModule, SharedModule],
+  imports: [TaggingInterfaceRoutingModule, SharedModule, VideoRecordingModule],
   providers: [],
 })
 export class TaggingInterfaceModule {}