+++ server/src/bucket/local.service.ts
@@ -1,5 +1,5 @@
 import { BucketStorage, BucketFile } from './bucket.service';
-import { copyFile, access, unlink } from 'fs/promises';
+import { copyFile, access, unlink, writeFile } from 'fs/promises';
 import { join, basename } from 'path';
 
 /**
@@ -16,13 +16,23 @@ export class LocalStorage extends BucketStorage {
     super(bucketName);
   }
 
-  async objectUpload(path: string, target: string): Promise<BucketFile> {
-    await copyFile(path, join(this.folder, target));
+  async objectUpload(
+    file: string | Buffer,
+    target: string,
+  ): Promise<BucketFile> {
+    // Replace '/' with '-' to avoid issues with folders
+    target = target.replace(/\//g, '-');
+
+    if (file instanceof Buffer) {
+      await writeFile(join(this.folder, target), file, {});
+    } else {
+      await copyFile(file, join(this.folder, target));
+    }
     return { name: target, uri: `/media/${basename(target)}` };
   }
 
   objectDownload(path: string, target: string): Promise<void> {
-    return copyFile(join(this.folder, target), path);
+    return copyFile(join('Entries', this.folder, target), path);
   }
 
   objectExists(target: string): Promise<boolean> {