+++ server/src/tag/tag.module.ts
@@ -8,7 +8,10 @@ import { TagGuard } from './tag.guard';
 import { Tag, TagSchema } from './tag.schema';
 import { TagService } from './tag.service';
 import { EntryStudyModule } from '../entrystudy/entrystudy.module';
-import { SharedModule } from 'src/shared/shared.module';
+import { SharedModule } from '../shared/shared.module';
+import { BucketModule } from '../bucket/bucket.module';
+import { DatasetModule } from '../dataset/dataset.module';
+import { EntryModule } from '../entry/entry.module';
 
 @Module({
   imports: [
@@ -18,6 +21,9 @@ import { SharedModule } from 'src/shared/shared.module';
     forwardRef(() => UserStudyModule),
     forwardRef(() => EntryStudyModule),
     forwardRef(() => SharedModule),
+    forwardRef(() => BucketModule),
+    forwardRef(() => DatasetModule),
+    forwardRef(() => EntryModule),
   ],
   controllers: [TagController],
   providers: [TagService, TagGuard],